+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 4.2.1
+
+- [FIX] temporary remove `PointerDeviceKind.trackpad`
+- [FIX] fix `'double?'` type
+
+# 4.2.0
+
+- [Add] `enlargeFactor` option
+- [Add] `CenterPageEnlargeStrategy.zoom` option
+- [Add] `animateToClosest` option
+
+- [FIX] clear timer if widget was unmounted
+- [FIX] scroll carousel using touchpad
+
 # 4.1.1
 
 - Fix code formatting